fix build and hidragtion error, fix auth
