+++ src/app/api/serve/route.tsx
@@ -1,11 +1,11 @@
 import { cache } from '@/session/server-sessions';
-import { NextRequest, NextResponse } from 'next/server';
+import { NextResponse } from 'next/server';
 
-export async function GET(req: NextRequest, res: NextResponse) {
+export async function GET() {
   return NextResponse.json({ cache });
 }
 
-// try {
+// // try {
 //addDoc not workig
 //   // const docRef = await addDoc(collection(firestoreDatabase, 'users'), {
 //   //   email: dataForSession.email,
