+++ src/modules/UI/CircularLoader.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import * as React from 'react';
 import CircularProgress, {
   CircularProgressProps,
