+++ src/modules/common/button.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { CircularProgress } from '@mui/material';
 
 export type ButtonType = {
