+++ src/modules/common/input.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
 
 export const Input = ({ value, ...props }: InputProps) => {
