+++ src/session/client-session.tsx
@@ -12,13 +12,18 @@ export const setCookie = (name: string, value: string, days: number) => {
 };
 
 export const getCookie = (name: string) => {
+  // console.log('name', name);
   const nameEQ = name + '=';
+
+  // if (typeof document !== 'undefined') {
   const ca = document.cookie.split(';');
   for (let i = 0; i < ca.length; i++) {
     let c = ca[i];
     while (c.charAt(0) === ' ') c = c.substring(1, c.length);
     if (c.indexOf(nameEQ) === 0) return c.substring(nameEQ.length, c.length);
   }
+  // }
+
   return null;
 };
